fix: close mobile nav after redirect
